+++ solution/parser/lib/ecfr/structure.go
