+++ solution/parser/lib/ecfr/structure_test.go
