+++ solution/parser/lib/eregs/config.go
