+++ solution/parser/lib/eregs/config_test.go
