+++ solution/parser/lib/eregs/eregs.go
@@ -1,0 +1,53 @@
+package eregs
+
+import (
+	"os"
+	"net/url"
+	"path"
+	"context"
+	"time"
+
+	"github.com/cmsgov/cmcs-eregulations/lib/network"
+)
+
+// BaseURL is the URL of the eRegs service that will accept the post requests
+var BaseURL string
+
+var parserResultURL = "/ecfr_parser_result/%d"
+
+var postAuth = &network.PostAuth{
+	Username: os.Getenv("EREGS_USERNAME"),
+	Password: os.Getenv("EREGS_PASSWORD"),
+}
+
+func parseURL(apiPath string) (*url.URL, error) {
+	u, err := url.Parse(BaseURL)
+	if err != nil {
+		return nil, err
+	}
+	u.Path = path.Join(u.Path, apiPath)
+	return u, nil
+}
+
+// ParserResult is the struct used to send results to the eRegs server
+type ParserResult struct {
+	Title           int    `json:"title,string"`
+	Start           string `json:"start"`
+	End             string `json:"end"`
+	Workers         int    `json:"workers,string"`
+	Parts           string `json:"parts"`
+	Subchapters     string `json:"subchapters"`
+	SkippedVersions int    `json:"skippedVersions,string"`
+	TotalVersions   int    `json:"totalVersions,string"`
+	Errors          int    `json:"errors,string"`
+}
+
+// PostParserResult is the function that sends a parser result to the eRegs server
+func PostParserResult(ctx context.Context, p *ParserResult) (int, error) {
+	u, err := parseURL(fmt.Sprintf(parserResultURL, p.Title))
+	if err != nil {
+		return -1, err
+	}
+	p.End = time.Now().Format(time.RFC3339)
+	return network.SendJSON(ctx, u, p, true, postAuth, network.HTTPPost)
+}
