+++ solution/parser/lib/eregs/resources_test.go
